Use lock file to enforce emacs-overlay version lock
Previously, the emacs-overlay version (which determinines which
version of emacs and its packages is used) was locked by hardcoding
its version in both flake.nix and condifured-emacs.nix.

By using the flake instead of the configured-emacs.nix file in the
build command, the configuration now uses the lock file as a single
point of truth for enforcing this version.